remove trait in coroutine
