+++ open-coroutine-core/src/coroutine/korosensei.rs
@@ -3,7 +3,7 @@ use crate::common::{Current, Named};
 use crate::constants::CoroutineState;
 use crate::coroutine::local::CoroutineLocal;
 use crate::coroutine::suspender::Suspender;
-use crate::coroutine::{Coroutine, StateCoroutine};
+use crate::coroutine::Coroutine;
 use corosensei::stack::DefaultStack;
 use corosensei::trap::TrapHandlerRegs;
 use corosensei::{CoroutineResult, ScopedCoroutine};
@@ -31,7 +31,7 @@ where
 {
     name: String,
     inner: ScopedCoroutine<'c, Param, Yield, Result<Return, &'static str>, DefaultStack>,
-    state: Cell<CoroutineState<Yield, Return>>,
+    pub(crate) state: Cell<CoroutineState<Yield, Return>>,
     pub(crate) local: CoroutineLocal<'c>,
 }
 
@@ -306,6 +306,11 @@ where
             );
         }
     }
+
+    /// Returns the current state of this `StateCoroutine`.
+    pub fn state(&self) -> CoroutineState<Yield, Return> {
+        self.state.get()
+    }
 }
 
 impl<Param, Yield, Return> Drop for CoroutineImpl<'_, Param, Yield, Return>
@@ -431,23 +436,5 @@ where
                 Ok(r)
             }
         }
-    }
-}
-
-impl<'c, Param, Yield, Return> StateCoroutine<'c> for CoroutineImpl<'c, Param, Yield, Return>
-where
-    Param: UnwindSafe,
-    Yield: Copy + Eq + PartialEq + UnwindSafe + Debug,
-    Return: Copy + Eq + PartialEq + UnwindSafe + Debug,
-{
-    fn state(&self) -> CoroutineState<Self::Yield, Self::Return> {
-        self.state.get()
-    }
-
-    fn change_state(
-        &self,
-        state: CoroutineState<Self::Yield, Self::Return>,
-    ) -> CoroutineState<Self::Yield, Self::Return> {
-        self.state.replace(state)
     }
 }
