+++ open-coroutine-core/src/coroutine/mod.rs
@@ -1,9 +1,8 @@
 use crate::common::{Current, Named};
-use crate::constants::{CoroutineState, Syscall, SyscallState};
+use crate::constants::CoroutineState;
 use crate::coroutine::suspender::Suspender;
 use crate::{impl_current_for, impl_display_by_debug, impl_for_named};
 use std::fmt::{Debug, Formatter};
-use std::io::{Error, ErrorKind};
 use std::ops::Deref;
 use std::panic::UnwindSafe;
 
@@ -12,6 +11,9 @@ pub mod suspender;
 
 /// Coroutine local abstraction.
 pub mod local;
+
+/// Coroutine state abstraction and impl.
+pub mod state;
 
 /// Create a new coroutine.
 #[macro_export]
@@ -102,207 +104,6 @@ impl<'c, SimpleCoroutineImpl: Coroutine<'c, Resume = ()>> SimpleCoroutine<'c>
     }
 }
 
-/// A trait implemented for describing changes in the state of the coroutine.
-pub trait StateCoroutine<'c>: Coroutine<'c>
-where
-    Self::Yield: Eq + PartialEq,
-    Self::Return: Eq + PartialEq,
-{
-    /// Returns the current state of this `StateCoroutine`.
-    fn state(&self) -> CoroutineState<Self::Yield, Self::Return>;
-
-    /// Returns the previous state of this `StateCoroutine`.
-    /// Note: user should not use this method.
-    fn change_state(
-        &self,
-        state: CoroutineState<Self::Yield, Self::Return>,
-    ) -> CoroutineState<Self::Yield, Self::Return>;
-
-    /// created -> ready
-    /// suspend -> ready
-    ///
-    /// # Errors
-    /// if change state fails.
-    fn ready(&self) -> std::io::Result<()> {
-        let current = self.state();
-        match current {
-            CoroutineState::Created => {
-                _ = self.change_state(CoroutineState::Ready);
-                return Ok(());
-            }
-            CoroutineState::Suspend(_, timestamp) => {
-                if timestamp <= open_coroutine_timer::now() {
-                    _ = self.change_state(CoroutineState::Ready);
-                    return Ok(());
-                }
-            }
-            _ => {}
-        }
-        Err(Error::new(
-            ErrorKind::Other,
-            format!(
-                "{} unexpected {current}->{:?}",
-                self.get_name(),
-                CoroutineState::<Self::Yield, Self::Return>::Ready
-            ),
-        ))
-    }
-
-    /// ready -> running
-    /// syscall -> running
-    ///
-    /// below just for test
-    /// created -> running
-    /// suspend -> running
-    ///
-    /// # Errors
-    /// if change state fails.
-    fn running(&self) -> std::io::Result<()> {
-        let current = self.state();
-        match current {
-            CoroutineState::Running => return Ok(()),
-            #[cfg(test)]
-            CoroutineState::Created => {
-                _ = self.change_state(CoroutineState::Running);
-                return Ok(());
-            }
-            CoroutineState::Ready => {
-                _ = self.change_state(CoroutineState::Running);
-                return Ok(());
-            }
-            // #[cfg(test)] preemptive.rs use this
-            CoroutineState::Suspend(_, timestamp) => {
-                if timestamp <= open_coroutine_timer::now() {
-                    _ = self.change_state(CoroutineState::Running);
-                    return Ok(());
-                }
-            }
-            CoroutineState::SystemCall(
-                _,
-                _,
-                SyscallState::Executing | SyscallState::Finished | SyscallState::Timeout,
-            ) => {
-                let state = CoroutineState::Running;
-                _ = self.change_state(state);
-                crate::info!("{} {:?}->{:?}", self.get_name(), current, state);
-                return Ok(());
-            }
-            _ => {}
-        }
-        Err(Error::new(
-            ErrorKind::Other,
-            format!(
-                "{} unexpected {current}->{:?}",
-                self.get_name(),
-                CoroutineState::<Self::Yield, Self::Return>::Running
-            ),
-        ))
-    }
-
-    /// running -> suspend
-    ///
-    /// # Errors
-    /// if change state fails.
-    fn suspend(&self, val: Self::Yield, timestamp: u64) -> std::io::Result<()> {
-        let current = self.state();
-        if CoroutineState::Running == current {
-            _ = self.change_state(CoroutineState::Suspend(val, timestamp));
-            return Ok(());
-        }
-        Err(Error::new(
-            ErrorKind::Other,
-            format!(
-                "{} unexpected {current}->{:?}",
-                self.get_name(),
-                CoroutineState::<Self::Yield, Self::Return>::Suspend(val, timestamp)
-            ),
-        ))
-    }
-
-    /// running -> syscall
-    /// inner: syscall -> syscall
-    ///
-    /// # Errors
-    /// if change state fails.
-    fn syscall(
-        &self,
-        val: Self::Yield,
-        syscall: Syscall,
-        syscall_state: SyscallState,
-    ) -> std::io::Result<()> {
-        let current = self.state();
-        match current {
-            CoroutineState::Running => {
-                let state = CoroutineState::SystemCall(val, syscall, syscall_state);
-                _ = self.change_state(state);
-                crate::info!("{} {:?}->{:?}", self.get_name(), current, state);
-                return Ok(());
-            }
-            CoroutineState::SystemCall(_, original_syscall, _) => {
-                if original_syscall == syscall {
-                    let state = CoroutineState::SystemCall(val, syscall, syscall_state);
-                    _ = self.change_state(state);
-                    crate::info!("{} {:?}->{:?}", self.get_name(), current, state);
-                    return Ok(());
-                }
-            }
-            _ => {}
-        }
-        Err(Error::new(
-            ErrorKind::Other,
-            format!(
-                "{} unexpected {current}->{:?}",
-                self.get_name(),
-                CoroutineState::<Self::Yield, Self::Return>::SystemCall(
-                    val,
-                    syscall,
-                    syscall_state
-                )
-            ),
-        ))
-    }
-
-    /// running -> complete
-    ///
-    /// # Errors
-    /// if change state fails.
-    fn complete(&self, val: Self::Return) -> std::io::Result<()> {
-        let current = self.state();
-        if CoroutineState::Running == current {
-            _ = self.change_state(CoroutineState::Complete(val));
-            return Ok(());
-        }
-        Err(Error::new(
-            ErrorKind::Other,
-            format!(
-                "{} unexpected {current}->{:?}",
-                self.get_name(),
-                CoroutineState::<Self::Yield, Self::Return>::Complete(val)
-            ),
-        ))
-    }
-
-    /// running -> error
-    ///
-    /// # Errors
-    /// if change state fails.
-    fn error(&self, val: &'static str) -> std::io::Result<()> {
-        let current = self.state();
-        if CoroutineState::Running == current {
-            _ = self.change_state(CoroutineState::Error(val));
-            return Ok(());
-        }
-        Err(Error::new(
-            ErrorKind::Other,
-            format!(
-                "{} unexpected {current}->{:?}",
-                self.get_name(),
-                CoroutineState::<Self::Yield, Self::Return>::Error(val)
-            ),
-        ))
-    }
-}
-
 impl_current_for!(
     COROUTINE,
     CoroutineImpl<'c, Param: UnwindSafe, Yield: Copy + UnwindSafe, Return: Copy + UnwindSafe>
@@ -311,8 +112,8 @@ impl_current_for!(
 impl<Param, Yield, Return> Debug for CoroutineImpl<'_, Param, Yield, Return>
 where
     Param: UnwindSafe,
-    Yield: Copy + Eq + PartialEq + UnwindSafe + Debug,
-    Return: Copy + Eq + PartialEq + UnwindSafe + Debug,
+    Yield: Copy + UnwindSafe + Debug,
+    Return: Copy + UnwindSafe + Debug,
 {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("Coroutine")
